added exit command and a ,essage for unknown commands

+++ lib/models/main.py
@@ -27,9 +27,16 @@ def main():
         elif command == "list":
             list_goals()
 
-        elif command == "complete" and cmd[1].isdigit():
-                complete_goal(int(cmd[1]))
+        elif command == "complete" and len(cmd) > 1 and cmd[1].isdigit():
+            complete_goal(int(cmd[1]))
 
-        elif command == "delete" and cmd[1].isdigit():
-                delete_goal(int(cmd[1]))        
+        elif command == "delete" and len(cmd) > 1 and cmd[1].isdigit():
+            delete_goal(int(cmd[1]))      
+
+        elif command == "exit":
+            break
+
+        else:
+            print("Unknown command.")
+            print_help()
